requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
kafka-python
python-dotenv
